requirements.txt
deepface==0.0.93
tf-keras==2.16.0
Pillow==10.3.0
numpy==1.24.4
pandas==2.0.3
scikit-learn==1.3.2
xgboost==2.0.3
shap==0.44.1
streamlit==1.32.2
joblib==1.3.2
matplotlib==3.8.4
setuptools==70.3.0
requests==2.33.0
